Alright good nuff

+++ bootstrap_mac.sh
@@ -4,9 +4,7 @@ set -u
 
 cd "$(dirname $0)" || exit
 
-# I don't do xcode-select install in this script because it is needed for git, so I can't get this script without it
-#  comment is here so I remember that next time
-# xcode-select install
+$(pwd -P)/git_setup.sh
 
 if [ -x "$(command -v brew)" ]; then
   brew update
@@ -25,7 +23,6 @@ export LDFLAGS="-L$(brew --prefix)/opt/readline/lib -L$(brew --prefix)/opt/zlib/
 
 brew upgrade
 
-# brew install asdf # Currently sucks at picking the right bundler version
 brew install awscli
 brew install coreutils
 brew install freetype
@@ -41,29 +38,24 @@ brew install nvm
 brew install openssl
 brew install postgresql
 brew install redis
-# brew install rust  ## Currently hangs
 brew install shellcheck
 brew install sqlite
 brew install terraform
 brew install vim
 brew install zsh
 
-brew install --cask chef-workstation
+git clone https://github.com/asdf-vm/asdf.git ~/.asdf --branch v0.14.0
+. "$HOME/.asdf/asdf.sh"
+asdf update
+
+curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh -s -- -y
+
 brew install --cask chromedriver
 brew install --cask firefox
-brew install --cask iterm2
 brew install --cask libreoffice
 brew install --cask spotify
 brew install --cask spotmenu
-brew install --cask virtualbox
 
-if [ ! -e ~/.vimrc ]; then
-  cp "$(pwd -P)/.vimrc" ~/.vimrc
-fi
-
-if [ ! -e ~/Library/Application\ Support/iTerm2/DynamicProfiles/iterm_profiles.json ]; then
-  cp "$(pwd -P)/iterm_profiles.json" ~/Library/Application\ Support/iTerm2/DynamicProfiles/iterm_profiles.json
-fi
 
 # I have to look into my license with this
 # if [ ! -d "/Applications/Skitch.app" ]; then
@@ -93,16 +85,12 @@ function installdmg {
 
 # This failed installing -- the download worked fine though
 # if [ ! -d "/Applications/Rancher Desktop.app" ]; then
-#  installdmg https://github.com/rancher-sandbox/rancher-desktop/releases/download/v1.2.0/Rancher.Desktop-1.2.0.aarch64.dmg
+#  installdmg https://github.com/rancher-sandbox/rancher-desktop/releases/download/v1.12.2/Rancher.Desktop-1.12.2.aarch64.dmg
 # fi
+
 if [ ! -e "/Applications/Google Chrome.app" ]; then
   installdmg https://dl.google.com/chrome/mac/universal/stable/GGRO/googlechrome.dmg
 fi
-
-# Failed to download
-# if [ ! -d "/Applications/zoom.us.app" ]; then
-#   sudo installer -pkg https://zoom.us/client/latest/Zoom.pkg -target /
-# fi
 
 if [ ! -e ~/.oh-my-zsh ]; then
   rm ~/.zshrc # Prevent oh my zsh from making a copy
@@ -113,12 +101,20 @@ if [ ! -e ~/.oh-my-zsh/themes/dracula/dracula.zsh-theme ]; then
   git clone git@github.com:dracula/zsh ~/.oh-my-zsh/themes/dracula
 fi
 
-cp "$(pwd -P)/.zshrc" ~/
+if [ ! -e ~/.pryrc ]; then
+  cp "$(pwd -P)/.pryrc" ~/
+fi
+
+if [ ! -e ~/.vimrc ]; then
+  cp "$(pwd -P)/.vimrc" ~/
+fi
+
+if [ ! -e ~/.zshrc ]; then
+  cp "$(pwd -P)/.zshrc" ~/
+fi
 
 # I don't think I care about this right now
 # cp -i "$(pwd -P)/.bash_profile.macos" ~/.bash_profile
-
-. /opt/homebrew/opt/asdf/libexec/asdf.sh
 
 # Needed sudo permissions here or chmod /usr/local/opt... not sure what i want to do
 asdf plugin add ruby
